reworked getNewMsgs()
In the previous version of the method, its second part (the for loop which only ever gets executed if there are more than 10 messages, so basically once at the start and very rarely afterwards) would incorrectly slice the message list, always counting the starting element as new regardless of whether this was true.

On top of fixing the above, the for loop is now a recursive lambda function that doesn't use any variables and has no side effects. Just how I like it.